requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
matplotlib>=3.7.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.0.0
PyPDF2>=3.0.0
